VOICE/VOICE_ASSISTANT/requirements.txt
Flask
wikipedia-api
pyjokes
pyautogui
requests
psutil
pillow

